refactor(library): 카테고리 UI 설정을 services에서 constants로 분리
